Remember to take files out of the project file when they are removed from teh project. :-) Some compilers don't like that. getprojinfo.cpp was removed from mapimg.pro
I better, but still not great, fix has been done for the line cache bug. IMGIO is now a class used for all IMG file IO. The cache can be cleared by calling clearCache() or when teh object is destroyed. We are employing this second method by making the IMGIO object local to the mapimg_reproject function. When that function completes (projection done or aborted) the cache will be completely destroyed.

We could also change the inputDataMap (cache) key from QString to a QPair<QString, QString> and use the text line number and filename as a pair. That way we could keep the cahce between projections and the lookup would still work properly. This would require creating the IMGIO object earlier and passing it to mapimg_reproject().