Added some preprocessor directives to print out a bunch of debug information and/or the final sum of the residuals.
Changed to #undef those preprocessor directives at the end of the file since they're in a header.

Added another stopping condition which kills the loop if the parameters have not changed for a hard-coded number of times; this helps to prevent lambda from getting too big and makes the debug output have less garbage.

Increased the initial value lambda in hopes that this would help with convergence.