Big time changes to how forward and inverse work.
forward and inverse are now non-virtual functions int the base projection
class that do al needed initialization and unit conversions before calling the
virtual functions _forward() and _inverse() which now only take the input coordinates as
arguments and place all their results in the corresponding class member variables,
forward() and inverse() are responsible for setting the pointer paremeters as appropriate.

This is currently untested.