Changed so that getNumRows and getNumCols return by value rather than by const-reference.
Added a big description of the class.

Changed so that pre- and post-conditions are in the doxygen @pre and @post
format.

Added doxygen //@{ and //@} grouping symbols to several groups of functions
so that documentation that applies to a group is duplicated for each member
of the group in the generated documentation.