Changed so that the getValue functions return by value, since they are just returning a value read from a file and aren't storing it in memory.
Added a bunch of get*() functions so that the user can access information that
was read from the header file.

Added some exception classes that are thrown when the user passes bad
parameters to the getValue functions.

Made the bsqFile ifstream mutable, so the user can treat it the object like
the class read it completely into memory and they're just accessing it from
there.  Also changed so that the bsqFile is opened as a binary file, because
that's what it is.

Now store the lower right x- and y-coordinates since they in the hdr file.