Changed to just read all of the BSQ file's contents into a std::vector upon construction.  Hopefully this will fix the read errors in Windows.
Added a readBSQFile() function to read the BSQ file into the std::vector.

Removed the copy constructor and assignment operator since the C++-generated
ones will now work properly.

Added an init() function to be used by the constructors.  It ensures that the
header file is read before the BSQ file, since the function to read the BSQ
file uses information that is obtained from the header file.